+++ controller/views/v1/events/schema.py
@@ -1,9 +1,5 @@
 from rest_framework import serializers
 
 
-class SampleResponseSerializer(serializers.Serializer):
-    text = serializers.CharField()
-
-
 class EventResponse(serializers.Serializer):
     id = serializers.IntegerField()
